requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Cloud Integration Dependencies (optional - install as needed)
# AWS
boto3>=1.26.0
awscli>=1.27.0

# Azure
azure-cli>=2.45.0
azure-monitor-query>=1.1.0
azure-identity>=1.12.0

# Google Cloud
google-cloud-monitoring>=2.11.0
google-cloud-logging>=3.2.0

# Real-time processing
websockets>=10.4
asyncio-mqtt>=0.11.0
